* Add splashScreen to project.
