+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,19 @@
         android:theme="@style/Theme.InventoryPilot"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".splash.MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.InventoryPilot">
+            android:theme="@style/Theme.InventoryPilot.Splash">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".login.presentation.ui.LoginActivity"
+            android:exported="true"
+            android:theme="@style/Theme.InventoryPilot" />
     </application>
 
 </manifest>
