--- app/src/main/java/softspark/com/inventorypilot/MainActivity.kt
@@ -1,17 +1,0 @@
-package softspark.com.inventorypilot
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import softspark.com.inventorypilot.databinding.MainActivityBinding
-
-class MainActivity : ComponentActivity() {
-
-    private lateinit var binding: MainActivityBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = MainActivityBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-    }
-}
